+++ extras/tests/Helpers/Arduino.h
@@ -1,0 +1,8 @@
+// ArduinoJson - arduinojson.org
+// Copyright Benoit Blanchon 2014-2020
+// MIT License
+
+#pragma once
+
+#include "api/Stream.h"
+#include "api/String.h"